pyproject.toml
[project]
name = "videodownloader"
version = "0.1.0"
description = ""
authors = [
    {name = "Frank1o3",email = "jahdy1o3@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13, <3.15"
dependencies = [
    "kivy (>=2.3.1,<3.0.0)",
    "yt-dlp (>=2025.12.8,<2026.0.0)",
    "mutagen (>=1.47.0,<2.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "kivymd (>=1.2.0,<2.0.0)",
    "pyjnius (>=1.7.0,<2.0.0)",
    "ffmpeg (>=1.4,<2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.13,<0.15.0)",
    "mypy (>=1.19.1,<2.0.0)"
]

[tool.poetry.scripts]
yt-downloader = "videodownloader.main:main"